+++ internal/dao/service.go
@@ -428,3 +428,20 @@ func (s *Service) processPopularCategory(_ context.Context) error {
 
 	return nil
 }
+
+func (s *Service) ProcessPopularityIndexUpdate(_ context.Context, id uuid.UUID, index float64) error {
+	existed, err := s.repo.GetByID(id)
+	if err != nil && !errors.Is(err, gorm.ErrRecordNotFound) {
+		return fmt.Errorf("handle: %w", err)
+	}
+
+	if existed != nil {
+		existed.PopularityIndex = index
+		err := s.repo.Update(*existed)
+		if err != nil {
+			return fmt.Errorf("update dao #%s: %w", id, err)
+		}
+	}
+
+	return nil
+}